backend/requirements.txt
Flask
langchain
langchain-community
langchain-text-splitters
langchain-chroma
langchain-openai
langgraph
chromadb
openai
pandas
pydantic
python-dotenv
serverless-wsgi
